feat(LCRT): extract the construction logic
Also includes a basic graph library as petgraph's index handling for its matrix graph makes it difficult to iterate over nodes reliably.

+++ lcrt/src/area_source.rs
@@ -494,7 +494,7 @@ impl ConstructionNode {
         crate::eta(
             self.availability,
             children,
-            self.interfering_neighbours + added_interfering_nodes,
+            f32::from(self.interfering_neighbours) + f32::from(added_interfering_nodes),
         )
     }
